Leaf indicator for multilevel leaves
When tree has leaves at different levels of tree, it's very difficult to keep track of leaf nodes at higher levels, since they're formatted as row headers, with even the rowExpand plus sign if showTreeExpandNoChildren is true.

Once showTreeExpandNoChildren is set to false, there's no way to know if you're on a child of a parent or on the next row at the same level as the parent.

This change adds info-circled where the plusSquared/minusSquared would normally sit, which in combination with the tree indentation shows the (i) symbol at the correct tree level for any node without children. A simpler plain circle without the 'i' would be better, but there were no set up classes for this.

Instead of this, ideally the row-header class would not automatically apply to every row with a set tree level, but rather only to those with child nodes: computing leaf nodes as only those where the treeLevel is not set is useless for trees with leaves at different levels.